Add files via upload
I regret to inform you that I am unable to submit the assignment as originally required. However, I have made an effort to create an approximate model of it using my knowledge of Dart and Flutter. Please note that I am still in the process of learning Kotlin and Java.